chore: resolve lint warnings
