+++ frontends/nextjs/src/components/ui/organisms/navigation/pagination/PaginationPrevious.tsx
@@ -2,8 +2,9 @@
 
 import { forwardRef } from 'react'
 
+import { PreviousIcon } from './paginationIcons'
 import { PaginationLink } from './PaginationLink'
-import { type PaginationLinkProps, PreviousIcon } from './paginationUtils'
+import { type PaginationLinkProps } from './paginationTypes'
 
 const PaginationPrevious = forwardRef<HTMLButtonElement, Omit<PaginationLinkProps, 'children'>>(
   (props, ref) => {
