+++ frontends/nextjs/src/components/ui/organisms/navigation/pagination/TablePagination.tsx
@@ -13,7 +13,7 @@ import {
 } from '@mui/material'
 import { forwardRef } from 'react'
 
-import { NextIcon, PreviousIcon } from './paginationUtils'
+import { NextIcon, PreviousIcon } from './paginationIcons'
 
 interface TablePaginationProps {
   count: number
