+++ frontends/nextjs/src/components/ui/organisms/navigation/pagination/index.ts
@@ -6,7 +6,7 @@ export { PaginationNext } from './PaginationNext'
 export { PaginationPrevious } from './PaginationPrevious'
 export type { PaginationProps } from './PaginationRoot'
 export { Pagination } from './PaginationRoot'
-export type { PaginationLinkProps } from './paginationUtils'
+export type { PaginationLinkProps } from './paginationTypes'
 export type { SimplePaginationProps } from './SimplePagination'
 export { SimplePagination } from './SimplePagination'
 export type { TablePaginationProps } from './TablePagination'
