+++ frontends/nextjs/src/components/ui/organisms/navigation/pagination/paginationIcons.tsx
@@ -1,0 +1,8 @@
+import ChevronLeftIcon from '@mui/icons-material/ChevronLeft'
+import ChevronRightIcon from '@mui/icons-material/ChevronRight'
+
+const PreviousIcon = () => <ChevronLeftIcon fontSize="small" />
+
+const NextIcon = () => <ChevronRightIcon fontSize="small" />
+
+export { NextIcon, PreviousIcon }
